fix: prevent daily production from dropping to 0
The API from the micro inverter sometimes drops the value for the day's production to 0 and after a small time back to its correct value. This commit introduces a caching mechanism to prevent these intermittent zero values from being propagated. The last valid production value is cached and returned when a zero value is received. The cache is reset at midnight to ensure the daily production value is correctly reset to 0 at the start of a new day.